Enabled to control the interval between clusters in latent space
Wae encodes the observation into latent space, which denotes xi.
The distribution of the latent varible makes up of clusters specified by
the external variable z.
It means that the variable xi with every z distribute in the different
cluster.
The parameter "cluster_interval" can control the interval between the
adjacent clusters