Added the pair of agent and environment which generate double rings
The pair of agent and environment for cs03a still generate the sample
from double rings and the two rings are conditionned by the independent
variable: z.
The pair for cs03b, of which the distribution follows the double rings,
is not conditional distribution, that is, the couple of rings is a
single distribution.